cli/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "spruce-cli"
version = "0.1.0"
description = "CLI for Sprucelab automation pipelines"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Sprucelab Team" }
]
keywords = ["bim", "ifc", "automation", "pipeline"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "typer>=0.9.0",
    "rich>=13.0.0",
    "httpx>=0.25.0",
    "pyyaml>=6.0",
    "keyring>=24.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]

[project.scripts]
spruce = "spruce.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["spruce"]
